third_party/WebKit/LayoutTests/platform/win/http/tests/security/mixedContent/insecure-picture-in-main-frame-blocked.https-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/insecure-picture-in-main-frame-blocked.https.html' was loaded over HTTPS, but requested an insecure image 'http://example.test:8080/security/resources/compass.jpg'. This request has been blocked; the content must be served over HTTPS.
This is a testharness.js-based test.
PASS Testing an insecure image in a <picture> element. 
Harness: the test ran to completion.
